+++ KTLog.m
@@ -481,12 +481,12 @@ static NSString *KTLevelMap[] = {
 #pragma mark -
 #pragma mark NSTableView Datasource
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
 	return [myLoggingLevels count];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
 	NSString *ident = [aTableColumn identifier];
 	NSDictionary *rec = [myLoggingLevels objectAtIndex:rowIndex];
@@ -500,7 +500,7 @@ static NSString *KTLevelMap[] = {
 	}
 }
 
-- (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
 	NSString *ident = [aTableColumn identifier];
 	NSMutableDictionary *rec = [myLoggingLevels objectAtIndex:rowIndex];
